fix(test): 修复测试功能异常
1. TEST_HEAD添加名字
2. gtest使用自定义的main函数
3. 修改ctest测试路径
