+++ test/test_pfring.cc
@@ -15,11 +15,7 @@
 #include <fmt/core.h>
 #include <helper.h>
 
-TEST_HEAD
+TEST_HEAD(test_pfring)
 {
-    struct lazybsd_version lazybsd_version = lazybsd_version_get();
-    fmt::print("version_major {} \nversion_minor {} \nversion_build {} \n",
-                lazybsd_version.version_major,
-                lazybsd_version.version_minor,
-                lazybsd_version.version_build);
+
 }
