+++ test/test_socket.cc
@@ -15,14 +15,7 @@
 #include <fmt/core.h>
 #include <helper.h>
 
-/**
- * @brief 测试原始socket实现
- *
- * @param argc
- * @param argv
- * @return int
- */
-int main(int argc, char *argv[])
+TEST_HEAD(test_socket)
 {
-    return 0;
+    
 }
